mas sobre validacion de entradas de datos en el back

+++ backend/controllers/controllersAdmin/inclusiveElements/deleteElement.js
@@ -1,10 +1,21 @@
 const Elements = require("../../../model/inclusiveElements.js");
 const cloudinary = require("../../../middlewares/cloudinary");
+const { _idMongooseRegex } = require("../../../regex") // Traemos los regex necesarios para validación de entradas
 
 const deleteElement = async(req, res) =>{
 
-    const {_id, ...resto} = req.body;
-    
+    const {_id } = req.body;
+    /* Sanitización entradas */
+    // 1) Validar el tipo de dato
+    if(typeof(_id) !== 'string') return res.status(422).json({ message: "Tipo de dato de _id no es válido" });
+
+    // 2) Validar si cumple con los caracteres permitidos 
+    const isValid_id = _idMongooseRegex.test(_id);
+
+    if(isValid_id == false) return res.json({ message: "Formato de _id no es válido" }); // Caso malo
+    /* Fin sanitización entradas */
+
+
     await Elements.deleteOne({_id:_id})
     .then((element)=>{
         cloudinary.uploader.destroy(`inclusiveElements/${resto.name}`);
